port from mysql to mysqli extension
The vanilla mysql module in PHP has been deprecated for some time. The
warning messages about it's use are nearly militant now and I just ran
across a server that didn't even have the old module to fall back on.
This simple switch should make use of the mysqli module instead with is
the closest drop-in replacement.