remove manual over-ride of table engine
The SQL syntax here was not valid and choked newer mysql family database
servers. MariaDB for example does not take this syntax. You can use
`engine=MyISAM` instead, but I think it's really better to leave this to
the default set for the parent database as some servers are optimized
for specific storage engines.