Remove the $die_on_error constructor argument
In favour of throwing exceptions where there are errors. This
library shouldn't ever die (the redirection added in this patch
is just a temporary workaround that will be removed in the next
breaking version).

Also switch to Travis, so we can include private env vars for the
test API consumer key.